Decimal thing added

+++ tables/tables.sql
@@ -28,11 +28,11 @@ CREATE TABLE Partner (
 );
 
 CREATE TABLE HealthCarePlan (
-  planName     VARCHAR(30) NOT NULL,
-  monthlyFee   INT         NOT NULL,
-  checkUp      INT         NOT NULL,
-  hygieneVisit INT         NOT NULL,
-  repairWork   INT         NOT NULL,
+  planName     VARCHAR(30)    NOT NULL,
+  monthlyFee   DECIMAL(5, 2)  NOT NULL,
+  checkUp      INT            NOT NULL,
+  hygieneVisit INT            NOT NULL,
+  repairWork   INT            NOT NULL,
   PRIMARY KEY (planName)
 );
 
@@ -61,9 +61,9 @@ CREATE TABLE Appointment (
 );
 
 CREATE TABLE Treatment (
-  name VARCHAR(30) NOT NULL,
-  type VARCHAR(20) NOT NULL,
-  cost INT         NOT NULL,
+  name VARCHAR(30)    NOT NULL,
+  type VARCHAR(20)    NOT NULL,
+  cost DECIMAL(5, 2)  NOT NULL,
   PRIMARY KEY (name)
 );
